+++ examples/scatter_and_line_plot_examples.rs
@@ -32,6 +32,8 @@ fn line_and_scatter_plot() {
     plot.add_trace(trace1);
     plot.add_trace(trace2);
     plot.add_trace(trace3);
+    plot.show_png(1024, 680);
+    plot.show_jpg(1024, 680);
     plot.show();
 }